docetressabores: Fix route paths and return type
Route paths were finishing with '/all' when using a GET transaction to get all available
content. The following will be the path convention from now on:

| Request | Example with users  | Action                       |
| :------ | :------------------ | :--------------------------- |
| GET     | /path/to/users      | gets all users               |
| GET     | /path/to/users/{id} | gets user with id = {id}     |
| POST    | /path/to/users      | creates user                 |
| DELETE  | /path/to/users/{id} | deletes users with id = {id} |
| PUT     | /path/to/users/{id} | updates user with id = {id}  |

The respective return types were plain data objects and the methods were not safe as
an exception could occurr if the wrong data were to be passed. These functions are
now safe and the return type was fixed to Response, a class which returns the response
status header as well as a body with the object/s if needed.

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>